Implement brand-new main UI
